demo_actor/requirements.txt
dapr-ext-fastapi==1.8.3